# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Active Home Platform API
  version: 1.0.0
  summary: Configuration-driven housing & construction advertising marketplace (Uzbekistan) — v1 contract.
  description: "Enterprise OpenAPI 3.1 contract for **Active Home v1**, the implementation contract between\
    \ the\nweb frontend, backend (FastAPI modular monolith), AI coding agents, QA automation, future mobile\n\
    apps, and external integrations.\n\nThis document is generated from and fully aligned with the immutable\
    \ approved architecture:\nSRS v1.0, DDD Domain Model v1.0, Configuration & Metadata Framework v1.0,\
    \ SAD v1.0,\nDatabase Architecture & Logical Data Model v1.0, Physical Database Design v1.0, and the\n\
    API Architecture & Integration Specification v1.0.\n\n## Cross-cutting contract rules (see the API\
    \ Architecture Specification)\n- **Versioning:** URI major versioning under `/api/v1`. Within a major\
    \ version only additive,\n  backward-compatible change is allowed. Configuration changes (new categories,\
    \ forms, fields,\n  plans, roles, templates) are **never** contract changes — dynamic data flows through\
    \ stable\n  envelopes (`AttributeMap`, `LocalizedText`, `FormDefinition`), so the endpoint surface\
    \ is\n  invariant under the platform's normal mode of change.\n- **Authentication:** server-side sessions\
    \ backed by Redis (SAD §13). A session is established by an\n  auth flow and carried by a secure,\
...
  contact:
...
  license:
...
servers:
...
security:
- sessionCookie: []
...
paths:
...
  /business-profiles:
    get:
      tags:
      - Business Profiles
      operationId: listBusinessProfiles
      summary: List business profiles
      description: Public directory of active business profiles; supports cursor pagination and filtering
        by type.
      security: []
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: profileType
        in: query
        schema:
          type: string
          enum:
          - CONSTRUCTION_COMPANY
          - MANUFACTURER
          - BUILDER
          - SUPPLIER
          - CONTRACTOR
          - ARCHITECT
          - INTERIOR_DESIGNER
          - SERVICE_PROVIDER
      - name: verifiedOnly
        in: query
        schema:
          type: boolean
      responses:
        '200':
          description: A page of business profiles.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/CursorPage'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/BusinessProfile'
    post:
      tags:
      - Business Profiles
      operationId: createBusinessProfile
      summary: Create a company profile
      description: Creates a business profile owned by the authenticated account (FR-PROF-001). One account
        may own many.
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessProfileCreateRequest'
      responses:
        '201':
          description: Created profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /business-profiles/{profileId}:
    get:
      tags:
      - Business Profiles
      operationId: getBusinessProfile
      summary: Get a business profile
      security: []
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '200':
          description: The profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessProfile'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Business Profiles
      operationId: updateBusinessProfile
      summary: Manage a business profile
      description: Updates a profile the caller owns (FR-PROF-002). `profileType` is immutable. Ownership
        is validated.
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessProfileUpdateRequest'
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - Business Profiles
      operationId: archiveBusinessProfile
      summary: Archive a business profile
      description: Archives the profile (state, not deletion — DM-06). Referenced listings/orders remain
        valid.
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '204':
          description: Archived.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /business-profiles/{profileId}/portfolio:
    get:
      tags:
      - Business Profiles
      operationId: listPortfolio
      summary: List portfolio items
      security: []
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '200':
          description: Ordered portfolio.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PortfolioItem'
    post:
      tags:
      - Business Profiles
      operationId: addPortfolioItem
      summary: Add a portfolio item
      description: Attaches an image (MediaAssetRef) to the profile portfolio.
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioItem'
      responses:
        '201':
          description: Added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioItem'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /business-profiles/{profileId}/portfolio/{itemId}:
    delete:
      tags:
      - Business Profiles
      operationId: removePortfolioItem
      summary: Remove a portfolio item
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      - name: itemId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Removed.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /business-profiles/{profileId}/team:
    get:
      tags:
      - Business Profiles
      operationId: listTeamMembers
      summary: List team members
      description: 'Lists users granted a role on this profile. Realised as identity role-assignments
        scoped to the

        profile — no new aggregate (see Validation note). Requires manage-team permission.

        '
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '200':
          description: Team members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMember'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
      - Business Profiles
      operationId: addTeamMember
      summary: Add a team member
      description: Grants a configured role to a user on this profile (ownership + permission validated).
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMember'
      responses:
        '201':
          description: Added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /business-profiles/{profileId}/team/{userId}:
    delete:
      tags:
      - Business Profiles
      operationId: removeTeamMember
      summary: Remove a team member
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Removed.
        '403':
          $ref: '#/components/responses/Forbidden'
  /business-profiles/{profileId}/verification:
    get:
      tags:
      - Business Profiles
      operationId: getVerificationCase
      summary: Get the current verification case
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      responses:
        '200':
          description: Verification case.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationCase'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags:
      - Business Profiles
      operationId: requestVerification
      summary: Request business verification
      description: 'Submits a verification request with documents. Requires an active VerificationEligibility

        entitlement (I-12, eventually consistent — a brief window may exist after payment confirmation).

        '
      parameters:
      - $ref: '#/components/parameters/ProfileIdPath'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationRequestCreate'
      responses:
        '201':
          description: Verification case created and queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationCase'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
  /admin/verification-queue:
    get:
      tags:
      - Administration
      operationId: listVerificationQueue
      summary: List verification cases
      description: The manual verification reviewer queue with SLA (DEC-13).
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - REQUESTED
          - IN_REVIEW
          - APPROVED
          - REJECTED
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of verification cases.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/CursorPage'
                - type: object
                  properties:
                    items:
                      type: array
                      items:
                        $ref: '#/components/schemas/VerificationCase'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/verification-queue/{caseId}/decision:
    post:
      tags:
      - Administration
      operationId: decideVerification
      summary: Decide a verification case
      description: Approves or rejects a verification case (FR-PROF-005). Approval issues the badge with
        validity terms (I-13).
      parameters:
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationDecisionRequest'
      responses:
        '200':
          description: Decision recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerificationCase'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ah_session
      description: 'First-party, server-side session (Redis-backed, SAD §13). Issued by any auth flow
        and sent as a

        secure, HttpOnly, SameSite=Lax cookie. This is the default scheme for browser clients.

        '
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 'The same first-party session token presented as `Authorization: Bearer <token>` for
        native and

        mobile clients that cannot use cookies. The token is an **opaque server-side session reference**,

        not a JWT. Revoking the session (logout) invalidates it immediately.

        '
    googleOAuth:
      type: oauth2
      description: 'Google is a federated **sign-in** identity provider only (authorization-code flow).
        After Google

        authenticates the user, Active Home links or creates the account and issues its own first-party

        session (`sessionCookie`/`sessionBearer`). Active Home does not expose an OAuth2 authorization

        server of its own.

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: OpenID Connect sign-in
...
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque forward cursor from a previous page's `nextCursor`. Omit for the first page.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      description: Page size. Values above the maximum are clamped, not rejected.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 'Client-generated key (UUID recommended) making a side-effecting POST safe to retry.
        The first

        successful result is replayed for subsequent identical requests.

        '
      schema:
        type: string
        format: uuid
...
    ProfileIdPath:
      name: profileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Unauthorized:
      description: No valid session, or session expired/revoked.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            noSession:
              $ref: '#/components/examples/ErrUnauthorized'
    Forbidden:
      description: Authenticated but not permitted (default-deny; wrong acting profile; insufficient permission
        key).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            denied:
              $ref: '#/components/examples/ErrForbidden'
    NotFound:
      description: Resource does not exist or is not visible to the caller (opaque).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            notFound:
              $ref: '#/components/examples/ErrNotFound'
...
    UnprocessableEntity:
      description: Business-rule violation or field-level validation failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              $ref: '#/components/examples/ErrValidation'
            businessRule:
              $ref: '#/components/examples/ErrBusinessRule'
...
  schemas:
...
    CursorPage:
      type: object
      description: Cursor pagination envelope (opaque forward cursor).
      required:
      - items
      - page
      properties:
        items:
          type: array
          items: {}
        page:
          type: object
          required:
          - limit
          properties:
            limit:
              type: integer
            nextCursor:
              type:
              - string
              - 'null'
              description: Pass as `cursor` to fetch the next page; null when exhausted.
            total:
              type:
              - integer
              - 'null'
              description: Present only where cheap to compute; may be null.
      additionalProperties: false
...
    BusinessProfile:
      type: object
      description: A company presence of one of the eight closed profile types (BC-02).
      required:
      - id
      - ownerUserId
      - profileType
      - name
      - status
      - slug
      - createdAt
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        ownerUserId:
          type: string
          format: uuid
          readOnly: true
        profileType:
          type: string
          description: Closed platform vocabulary (eight types).
          enum:
          - CONSTRUCTION_COMPANY
          - MANUFACTURER
          - BUILDER
          - SUPPLIER
          - CONTRACTOR
          - ARCHITECT
          - INTERIOR_DESIGNER
          - SERVICE_PROVIDER
        name:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        contacts:
          type: object
          additionalProperties: true
          description: Phones/emails/site (CompanyDetails).
        address:
          type:
          - string
          - 'null'
        slug:
          type: string
          readOnly: true
          description: Display-only routing; not an identity.
        status:
          type: string
          enum:
          - CREATED
          - ACTIVE
          - ARCHIVED
          readOnly: true
        badge:
          type:
          - object
          - 'null'
          readOnly: true
          description: Trust badge sub-state (null = never verified).
          properties:
            status:
              type: string
              enum:
              - VALID
              - EXPIRED
              - REVOKED
            issuedAt:
              type: string
              format: date-time
            validUntil:
              type: string
              format: date-time
          additionalProperties: false
        createdAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
    BusinessProfileCreateRequest:
      type: object
      required:
      - profileType
      - name
      properties:
        profileType:
          type: string
          enum:
          - CONSTRUCTION_COMPANY
          - MANUFACTURER
          - BUILDER
          - SUPPLIER
          - CONTRACTOR
          - ARCHITECT
          - INTERIOR_DESIGNER
          - SERVICE_PROVIDER
        name:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        contacts:
          type: object
          additionalProperties: true
        address:
          type: string
      additionalProperties: false
    BusinessProfileUpdateRequest:
      type: object
      description: Partial update; profileType is immutable and cannot be changed.
      properties:
        name:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        contacts:
          type: object
          additionalProperties: true
        address:
          type: string
      additionalProperties: false
    PortfolioItem:
      type: object
      required:
      - id
      - mediaAssetId
      - position
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        mediaAssetId:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 1
          maximum: 50
        caption:
          $ref: '#/components/schemas/LocalizedText'
      additionalProperties: false
    TeamMember:
      type: object
      description: 'A user granted a role on a business profile (realised as an identity RoleAssignment
        scoped to the

        profile; the profile owner manages membership). No new aggregate is introduced.

        '
      required:
      - userId
      - roleCode
      properties:
        userId:
          type: string
          format: uuid
        displayName:
          type:
          - string
          - 'null'
          readOnly: true
        roleCode:
          type: string
          description: A configured role code.
        invitedAt:
          type: string
          format: date-time
          readOnly: true
      additionalProperties: false
...